Center variable

Center a variable by subtracting the mean from each element. Centering can be performed by the grand mean when by = NULL (the default), or by group means when by is a factor variable.

center_variable(variable, scale = FALSE, by = NULL)

Arguments

variable

A numeric vector.

scale

A logical vector. If scale = TRUE, the centered values of variable are divided by their standard deviation.

by

A vector with the factor class.

Value

A numeric vector.

Examples

df <- data.frame(
  id = 1:20,
  group = as.factor(sample(c("A", "B"), 20, replace = TRUE)),
  body_mass = rnorm(20, mean = 65, sd = 12)
)

df$body_mass_centered <- center_variable(df$body_mass, by = df$group)
df
#>    id group body_mass body_mass_centered
#> 1   1     A  62.25906       -12.67559515
#> 2   2     A  73.83027        -1.10439203
#> 3   3     A  61.12390       -13.81075406
#> 4   4     B  75.15272         3.15849728
#> 5   5     A  76.09354         1.15888391
#> 6   6     A  72.29583        -2.63882439
#> 7   7     B  71.97550        -0.01872769
#> 8   8     B 104.23124        32.23701206
#> 9   9     A  68.70692        -6.22773970
#> 10 10     B  69.16147        -2.83275308
#> 11 11     A  74.46079        -0.47386311
#> 12 12     B  89.67447        17.68024974
#> 13 13     B  61.43830       -10.55592590
#> 14 14     A  67.01607        -7.91858651
#> 15 15     A  86.49757        11.56291169
#> 16 16     A  92.16641        17.23174855
#> 17 17     B  46.81291       -25.18130964
#> 18 18     A  77.33334         2.39867871
#> 19 19     B  57.50718       -14.48704277
#> 20 20     A  87.43219        12.49753210