Leave-one-out cross-validation

Cross-validates the model using the leave-one-out approach. In this method each subject's data is separated into a testing data set, and all other subject's are kept in the training data set, with as many resamples as the number of subjects in the original data set. It computes the model's predicted value in the testing data set for each subject.

loocv(model, data, id, keep = "all")

# S3 method for default
loocv(model, data, id, keep = "all")

# S3 method for lm
loocv(model, data, id, keep = "all")

# S3 method for lmerMod
loocv(model, data, id, keep = "all")

Arguments

model
An object containing a model.

data
A data frame.

id
The bare (unquoted) name of the column which identifies subjects.

keep
A character string which controls which columns are present in the output. Can be one of three options:

"all": The default. Retain all variables in the original data frame plus the ".actual" and ".predicted" columns.

"used": Keeps only the "id" column of the original data frame, plus the ".actual" and ".predicted" columns.

"none": Returns just the ".actual" and `".predicted" columns.

Value

Returns an object of class loocv. A tibble containing the ".actual" and ".predicted" columns.

Examples

data <- tibble::tibble(
  id = rep(1:10, each = 2),
  x = rnorm(20, mean = 100, sd = 10),
  y = x * runif(20, min = 1, max = 1.2)
)

olr <- stats::lm(y ~ x, data)
lmm <- lme4::lmer(y ~ x + (1 | id), data)

loocv(olr, data, id)

#> Joining, by = "id"

#> # A tibble: 40 x 5
#>       id     x     y .actual .predicted
#>    <int> <dbl> <dbl>   <dbl>      <dbl>
#>  1     1  97.8 115.    115.       107. 
#>  2     1  85.1  96.7    96.7       93.4
#>  3     1  97.8 115.    115.       107. 
#>  4     1  85.1  96.7    96.7       93.4
#>  5     2 112.  128.    128.       121. 
#>  6     2  85.2  88.3    88.3       96.6
#>  7     2 112.  128.    128.       121. 
#>  8     2  85.2  88.3    88.3       96.6
#>  9     3  95.7  97.2    97.2      107. 
#> 10     3  89.5  90.8    90.8      101. 
#> # … with 30 more rows

loocv(lmm, data, id)

#> boundary (singular) fit: see ?isSingular

#> Joining, by = "id"

#> # A tibble: 40 x 5
#>       id     x     y .actual .predicted
#>    <int> <dbl> <dbl>   <dbl>      <dbl>
#>  1     1  97.8 115.    115.       107. 
#>  2     1  85.1  96.7    96.7       93.6
#>  3     1  97.8 115.    115.       107. 
#>  4     1  85.1  96.7    96.7       93.6
#>  5     2 112.  128.    128.       121. 
#>  6     2  85.2  88.3    88.3       96.9
#>  7     2 112.  128.    128.       121. 
#>  8     2  85.2  88.3    88.3       96.9
#>  9     3  95.7  97.2    97.2      107. 
#> 10     3  89.5  90.8    90.8      101. 
#> # … with 30 more rows