Print all rows of a data frame or tibble

Shortcut to print all rows of a data frame or tibble. Useful to inspect the whole tibble, as it prints by default only the first 20 rows.

pa(data)

Arguments

data

A data frame or tibble.

Value

Prints data and returns it invisibly.

See also

Examples

df <- dplyr::starwars
pa(df)
#> # A tibble: 87 × 14
#>    name       height   mass hair_…¹ skin_…² eye_c…³ birth…⁴ sex   gender homew…⁵
#>    <chr>       <int>  <dbl> <chr>   <chr>   <chr>     <dbl> <chr> <chr>  <chr>  
#>  1 Luke Skyw…    172   77   blond   fair    blue       19   male  mascu… Tatooi…
#>  2 C-3PO         167   75   NA      gold    yellow    112   none  mascu… Tatooi…
#>  3 R2-D2          96   32   NA      white,… red        33   none  mascu… Naboo  
#>  4 Darth Vad…    202  136   none    white   yellow     41.9 male  mascu… Tatooi…
#>  5 Leia Orga…    150   49   brown   light   brown      19   fema… femin… Aldera…
#>  6 Owen Lars     178  120   brown,… light   blue       52   male  mascu… Tatooi…
#>  7 Beru Whit…    165   75   brown   light   blue       47   fema… femin… Tatooi…
#>  8 R5-D4          97   32   NA      white,… red        NA   none  mascu… Tatooi…
#>  9 Biggs Dar…    183   84   black   light   brown      24   male  mascu… Tatooi…
#> 10 Obi-Wan K…    182   77   auburn… fair    blue-g…    57   male  mascu… Stewjon
#> 11 Anakin Sk…    188   84   blond   fair    blue       41.9 male  mascu… Tatooi…
#> 12 Wilhuff T…    180   NA   auburn… fair    blue       64   male  mascu… Eriadu 
#> 13 Chewbacca     228  112   brown   unknown blue      200   male  mascu… Kashyy…
#> 14 Han Solo      180   80   brown   fair    brown      29   male  mascu… Corell…
#> 15 Greedo        173   74   NA      green   black      44   male  mascu… Rodia  
#> 16 Jabba Des…    175 1358   NA      green-… orange    600   herm… mascu… Nal Hu…
#> 17 Wedge Ant…    170   77   brown   fair    hazel      21   male  mascu… Corell…
#> 18 Jek Tono …    180  110   brown   fair    blue       NA   male  mascu… Bestin…
#> 19 Yoda           66   17   white   green   brown     896   male  mascu… NA     
#> 20 Palpatine     170   75   grey    pale    yellow     82   male  mascu… Naboo  
#> 21 Boba Fett     183   78.2 black   fair    brown      31.5 male  mascu… Kamino 
#> 22 IG-88         200  140   none    metal   red        15   none  mascu… NA     
#> 23 Bossk         190  113   none    green   red        53   male  mascu… Trando…
#> 24 Lando Cal…    177   79   black   dark    brown      31   male  mascu… Socorro
#> 25 Lobot         175   79   none    light   blue       37   male  mascu… Bespin 
#> 26 Ackbar        180   83   none    brown … orange     41   male  mascu… Mon Ca…
#> 27 Mon Mothma    150   NA   auburn  fair    blue       48   fema… femin… Chandr…
#> 28 Arvel Cry…     NA   NA   brown   fair    brown      NA   male  mascu… NA     
#> 29 Wicket Sy…     88   20   brown   brown   brown       8   male  mascu… Endor  
#> 30 Nien Nunb     160   68   none    grey    black      NA   male  mascu… Sullust
#> 31 Qui-Gon J…    193   89   brown   fair    blue       92   male  mascu… NA     
#> 32 Nute Gunr…    191   90   none    mottle… red        NA   male  mascu… Cato N…
#> 33 Finis Val…    170   NA   blond   fair    blue       91   male  mascu… Corusc…
#> 34 Jar Jar B…    196   66   none    orange  orange     52   male  mascu… Naboo  
#> 35 Roos Tarp…    224   82   none    grey    orange     NA   male  mascu… Naboo  
#> 36 Rugor Nass    206   NA   none    green   orange     NA   male  mascu… Naboo  
#> 37 Ric Olié      183   NA   brown   fair    blue       NA   NA    NA     Naboo  
#> 38 Watto         137   NA   black   blue, … yellow     NA   male  mascu… Toydar…
#> 39 Sebulba       112   40   none    grey, … orange     NA   male  mascu… Malast…
#> 40 Quarsh Pa…    183   NA   black   dark    brown      62   NA    NA     Naboo  
#> 41 Shmi Skyw…    163   NA   black   fair    brown      72   fema… femin… Tatooi…
#> 42 Darth Maul    175   80   none    red     yellow     54   male  mascu… Dathom…
#> 43 Bib Fortu…    180   NA   none    pale    pink       NA   male  mascu… Ryloth 
#> 44 Ayla Secu…    178   55   none    blue    hazel      48   fema… femin… Ryloth 
#> 45 Dud Bolt       94   45   none    blue, … yellow     NA   male  mascu… Vulpter
#> 46 Gasgano       122   NA   none    white,… black      NA   male  mascu… Troiken
#> 47 Ben Quadi…    163   65   none    grey, … orange     NA   male  mascu… Tund   
#> 48 Mace Windu    188   84   none    dark    brown      72   male  mascu… Haruun…
#> 49 Ki-Adi-Mu…    198   82   white   pale    yellow     92   male  mascu… Cerea  
#> 50 Kit Fisto     196   87   none    green   black      NA   male  mascu… Glee A…
#> 51 Eeth Koth     171   NA   black   brown   brown      NA   male  mascu… Iridon…
#> 52 Adi Gallia    184   50   none    dark    blue       NA   fema… femin… Corusc…
#> 53 Saesee Ti…    188   NA   none    pale    orange     NA   male  mascu… Iktotch
#> 54 Yarael Po…    264   NA   none    white   yellow     NA   male  mascu… Quermia
#> 55 Plo Koon      188   80   none    orange  black      22   male  mascu… Dorin  
#> 56 Mas Amedda    196   NA   none    blue    blue       NA   male  mascu… Champa…
#> 57 Gregar Ty…    185   85   black   dark    brown      NA   male  mascu… Naboo  
#> 58 Cordé         157   NA   brown   light   brown      NA   fema… femin… Naboo  
#> 59 Cliegg La…    183   NA   brown   fair    blue       82   male  mascu… Tatooi…
#> 60 Poggle th…    183   80   none    green   yellow     NA   male  mascu… Geonos…
#> 61 Luminara …    170   56.2 black   yellow  blue       58   fema… femin… Mirial 
#> 62 Barriss O…    166   50   black   yellow  blue       40   fema… femin… Mirial 
#> 63 Dormé         165   NA   brown   light   brown      NA   fema… femin… Naboo  
#> 64 Dooku         193   80   white   fair    brown     102   male  mascu… Serenno
#> 65 Bail Pres…    191   NA   black   tan     brown      67   male  mascu… Aldera…
#> 66 Jango Fett    183   79   black   tan     brown      66   male  mascu… Concor…
#> 67 Zam Wesell    168   55   blonde  fair, … yellow     NA   fema… femin… Zolan  
#> 68 Dexter Je…    198  102   none    brown   yellow     NA   male  mascu… Ojom   
#> 69 Lama Su       229   88   none    grey    black      NA   male  mascu… Kamino 
#> 70 Taun We       213   NA   none    grey    black      NA   fema… femin… Kamino 
#> 71 Jocasta Nu    167   NA   white   fair    blue       NA   fema… femin… Corusc…
#> 72 Ratts Tye…     79   15   none    grey, … unknown    NA   male  mascu… Aleen …
#> 73 R4-P17         96   NA   none    silver… red, b…    NA   none  femin… NA     
#> 74 Wat Tambor    193   48   none    green,… unknown    NA   male  mascu… Skako  
#> 75 San Hill      191   NA   none    grey    gold       NA   male  mascu… Muunil…
#> 76 Shaak Ti      178   57   none    red, b… black      NA   fema… femin… Shili  
#> 77 Grievous      216  159   none    brown,… green,…    NA   male  mascu… Kalee  
#> 78 Tarfful       234  136   brown   brown   blue       NA   male  mascu… Kashyy…
#> 79 Raymus An…    188   79   brown   light   brown      NA   male  mascu… Aldera…
#> 80 Sly Moore     178   48   none    pale    white      NA   NA    NA     Umbara 
#> 81 Tion Medon    206   80   none    grey    black      NA   male  mascu… Utapau 
#> 82 Finn           NA   NA   black   dark    dark       NA   male  mascu… NA     
#> 83 Rey            NA   NA   brown   light   hazel      NA   fema… femin… NA     
#> 84 Poe Damer…     NA   NA   brown   light   brown      NA   male  mascu… NA     
#> 85 BB8            NA   NA   none    none    black      NA   none  mascu… NA     
#> 86 Captain P…     NA   NA   unknown unknown unknown    NA   NA    NA     NA     
#> 87 Padmé Ami…    165   45   brown   light   brown      46   fema… femin… Naboo  
#> # … with 4 more variables: species <chr>, films <list>, vehicles <list>,
#> #   starships <list>, and abbreviated variable names ¹​hair_color, ²​skin_color,
#> #   ³​eye_color, ⁴​birth_year, ⁵​homeworld